requirements.txt
# Vector DB
weaviate-client>=4.5,<5.0

# LLM + Embeddings (OpenAI)
openai>=1.50

# LLM agnostique (LiteLLM — supporte OpenAI, Mistral, Anthropic, Ollama…)
litellm>=1.40

# Agent RAG (LangGraph)
langgraph>=0.2
pydantic>=2.0

# Reranking (Appel API vers Infinity/TEI BGE)
requests>=2.31.0

# API REST
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
fastapi-users[sqlalchemy]>=13.0  # authentification JWT Bearer

# Base de données (SQLAlchemy async + Alembic)
sqlalchemy[asyncio]>=2.0
alembic>=1.13
aiosqlite>=0.20         # driver SQLite async (dev / tests)
asyncpg>=0.29           # driver PostgreSQL async (production)
psycopg2-binary>=2.9    # driver sync PostgreSQL requis par Alembic

# UI Streamlit (mode POC — optionnel en production)
streamlit>=1.32

# Config
python-dotenv>=1.0

# Logging
loguru>=0.7

# PDF text extraction (fallback si openingestion n'est pas installé)
pymupdf>=1.23

# Object storage (MinIO / S3-compatible) — visual grounding
minio>=7.2

# Queue de tâches asynchrones (ingestion industrielle)
celery[redis]>=5.3
redis>=5.0

#openingestion
openingestion[mineru]>=0.1.5

# connectors
playwright>=1.42
